+++ elasticapm/base.py
@@ -12,61 +12,23 @@ Large portions are
 from __future__ import absolute_import
 
 import datetime
+import inspect
 import logging
 import os
 import platform
 import socket
 import sys
-import threading
-import time
-import zlib
 from copy import deepcopy
 
 import elasticapm
 from elasticapm.conf import Config, constants
+from elasticapm.conf.constants import ERROR
 from elasticapm.traces import TransactionsStore, get_transaction
-from elasticapm.transport.base import TransportException
-from elasticapm.utils import compat, is_master_process
-from elasticapm.utils import json_encoder as json
-from elasticapm.utils import stacks, varmap
+from elasticapm.utils import compat, is_master_process, stacks, varmap
 from elasticapm.utils.encoding import keyword_field, shorten, transform
 from elasticapm.utils.module_import import import_string
 
 __all__ = ("Client",)
-
-
-class ClientState(object):
-    ONLINE = 1
-    ERROR = 0
-
-    def __init__(self):
-        self.status = self.ONLINE
-        self.last_check = None
-        self.retry_number = 0
-
-    def should_try(self):
-        if self.status == self.ONLINE:
-            return True
-
-        interval = min(self.retry_number, 6) ** 2
-
-        if time.time() - self.last_check > interval:
-            return True
-
-        return False
-
-    def set_fail(self):
-        self.status = self.ERROR
-        self.retry_number += 1
-        self.last_check = time.time()
-
-    def set_success(self):
-        self.status = self.ONLINE
-        self.last_check = None
-        self.retry_number = 0
-
-    def did_fail(self):
-        return self.status == self.ERROR
 
 
 class Client(object):
@@ -105,13 +67,10 @@ class Client(object):
         cls = self.__class__
         self.logger = logging.getLogger("%s.%s" % (cls.__module__, cls.__name__))
         self.error_logger = logging.getLogger("elasticapm.errors")
-        self.state = ClientState()
 
         self.transaction_store = None
         self.processors = []
         self.filter_exception_types_dict = {}
-        self._send_timer = None
-        self._transports = {}
         self._service_info = None
 
         self.config = Config(config, inline_dict=inline)
@@ -120,7 +79,25 @@ class Client(object):
                 self.error_logger.error(msg)
             self.config.disable_send = True
 
-        self._transport_class = import_string(self.config.transport_class)
+        headers = {
+            "Content-Type": "application/x-ndjson",
+            "Content-Encoding": "gzip",
+            "User-Agent": "elasticapm-python/%s" % elasticapm.VERSION,
+        }
+
+        if self.config.secret_token:
+            headers["Authorization"] = "Bearer %s" % self.config.secret_token
+        transport_kwargs = {
+            "metadata": self._build_metadata(),
+            "headers": headers,
+            "verify_server_cert": self.config.verify_server_cert,
+            "timeout": self.config.server_timeout,
+            "max_flush_time": self.config.api_request_time / 1000.0,
+            "max_buffer_size": self.config.api_request_size,
+        }
+        self._transport = import_string(self.config.transport_class)(
+            compat.urlparse.urljoin(self.config.server_url, constants.EVENTS_API_PATH), **transport_kwargs
+        )
 
         for exc_to_filter in self.config.filter_exception_types or []:
             exc_to_filter_type = exc_to_filter.split(".")[-1]
@@ -136,9 +113,12 @@ class Client(object):
         else:
             skip_modules = ("elasticapm.",)
 
-        def frames_collector_func():
-            return self._get_stack_info_for_trace(
-                stacks.iter_stack_frames(skip_top_modules=skip_modules),
+        self.transaction_store = TransactionsStore(
+            frames_collector_func=lambda: list(
+                stacks.iter_stack_frames(start_frame=inspect.currentframe(), skip_top_modules=skip_modules)
+            ),
+            frames_processing_func=lambda frames: self._get_stack_info_for_trace(
+                frames,
                 library_frame_context_lines=self.config.source_lines_span_library_frames,
                 in_app_frame_context_lines=self.config.source_lines_span_app_frames,
                 with_locals=self.config.collect_local_variables in ("all", "transactions"),
@@ -150,15 +130,11 @@ class Client(object):
                     ),
                     local_var,
                 ),
-            )
-
-        self.transaction_store = TransactionsStore(
-            frames_collector_func=frames_collector_func,
-            collect_frequency=self.config.flush_interval,
+            ),
+            queue_func=self.queue,
             sample_rate=self.config.transaction_sample_rate,
             max_spans=self.config.transaction_max_spans,
-            span_frames_min_duration=self.config.span_frames_min_duration_ms,
-            max_queue_size=self.config.max_queue_size,
+            span_frames_min_duration=self.config.span_frames_min_duration,
             ignore_patterns=self.config.transactions_ignore_patterns,
         )
         self.include_paths_re = stacks.get_path_regex(self.config.include_paths) if self.config.include_paths else None
@@ -180,9 +156,9 @@ class Client(object):
         )
 
         if data:
-            url = self.config.server_url + constants.ERROR_API_PATH
-            self.send(url, **data)
-            return data["errors"][0]["id"]
+            # queue data, and flush the queue if this is an unhandled exception
+            self.queue(ERROR, data, flush=not handled)
+            return data["id"]
 
     def capture_message(self, message=None, param_message=None, **kwargs):
         """
@@ -208,133 +184,29 @@ class Client(object):
         """
         return self.capture("Exception", exc_info=exc_info, handled=handled, **kwargs)
 
-    def send(self, url, **data):
-        """
-        Encodes and sends data to remote URL using configured transport
-        :param url: URL of endpoint
-        :param data: dictionary of data to send
-        """
-        if self.config.disable_send or self._filter_exception_type(data):
+    def queue(self, event_type, data, flush=False):
+        if self.config.disable_send:
             return
+        # Run the data through processors
+        for processor in self.processors:
+            if not hasattr(processor, "event_types") or event_type in processor.event_types:
+                data = processor(self, data)
+        if flush and is_master_process():
+            # don't flush in uWSGI master process to avoid ending up in an unpredictable threading state
+            flush = False
+        self._transport.queue(event_type, data, flush)
 
-        payload = self.encode(data)
-
-        headers = {
-            "Content-Type": "application/json",
-            "Content-Encoding": "deflate",
-            "User-Agent": "elasticapm-python/%s" % elasticapm.VERSION,
-        }
-
-        if self.config.secret_token:
-            headers["Authorization"] = "Bearer %s" % self.config.secret_token
-
-        if not self.state.should_try():
-            message = self._get_log_message(payload)
-            self.error_logger.error(message)
-            return
-        try:
-            self._send_remote(url=url, data=payload, headers=headers)
-        except Exception as e:
-            self.handle_transport_fail(exception=e)
-
-    def encode(self, data):
-        """
-        Serializes ``data`` into a raw string.
-        """
-        return zlib.compress(json.dumps(data).encode("utf8"))
-
-    def decode(self, data):
-        """
-        Unserializes a string, ``data``.
-        """
-        return json.loads(zlib.decompress(data).decode("utf8"))
-
-    def begin_transaction(self, transaction_type):
+    def begin_transaction(self, transaction_type, trace_parent=None):
         """Register the start of a transaction on the client
         """
-        return self.transaction_store.begin_transaction(transaction_type)
+        return self.transaction_store.begin_transaction(transaction_type, trace_parent=trace_parent)
 
     def end_transaction(self, name=None, result=""):
         transaction = self.transaction_store.end_transaction(result, name)
-        if self.transaction_store.should_collect():
-            self._collect_transactions()
-        if not self._send_timer:
-            # send first batch of data after config._wait_to_first_send
-            self._start_send_timer(timeout=min(self.config._wait_to_first_send, self.config.flush_interval))
         return transaction
 
     def close(self):
-        self._collect_transactions()
-        if self._send_timer:
-            self._stop_send_timer()
-        for url, transport in list(self._transports.items()):
-            transport.close()
-            self._transports.pop(url)
-
-    def handle_transport_success(self, **kwargs):
-        """
-        Success handler called by the transport
-        """
-        if kwargs.get("url"):
-            self.logger.info("Logged error at " + kwargs["url"])
-        self.state.set_success()
-
-    def handle_transport_fail(self, exception=None, **kwargs):
-        """
-        Failure handler called by the transport
-        """
-        if isinstance(exception, TransportException):
-            message = self._get_log_message(exception.data)
-            self.error_logger.error(exception.args[0])
-        else:
-            # stdlib exception
-            message = str(exception)
-        self.error_logger.error(
-            "Failed to submit message: %r", message, exc_info=getattr(exception, "print_trace", True)
-        )
-        self.state.set_fail()
-
-    def _collect_transactions(self):
-        self._stop_send_timer()
-        transactions = []
-        if self.transaction_store:
-            for transaction in self.transaction_store.get_all():
-                for processor in self.processors:
-                    transaction = processor(self, transaction)
-                transactions.append(transaction)
-        if not transactions:
-            return
-
-        data = self._build_msg({"transactions": transactions})
-
-        api_path = constants.TRANSACTIONS_API_PATH
-
-        self.send(self.config.server_url + api_path, **data)
-        self._start_send_timer()
-
-    def _start_send_timer(self, timeout=None):
-        timeout = timeout or self.config.flush_interval
-        self._send_timer = threading.Timer(timeout, self._collect_transactions)
-        self._send_timer.start()
-
-    def _stop_send_timer(self):
-        if self._send_timer and not self._send_timer == threading.current_thread():
-            self._send_timer.cancel()
-            if self._send_timer.is_alive():
-                self._send_timer.join()
-
-    def _send_remote(self, url, data, headers=None):
-        if headers is None:
-            headers = {}
-        parsed = compat.urlparse.urlparse(url)
-        transport = self._get_transport(parsed)
-        if transport.async_mode:
-            transport.send_async(
-                data, headers, success_callback=self.handle_transport_success, fail_callback=self.handle_transport_fail
-            )
-        else:
-            url = transport.send(data, headers, timeout=self.config.server_timeout)
-            self.handle_transport_success(url=url)
+        self._transport.close()
 
     def get_service_info(self):
         if self._service_info:
@@ -378,13 +250,12 @@ class Client(object):
             "platform": platform.system().lower(),
         }
 
-    def _build_msg(self, data=None, **kwargs):
-        data = data or {}
-        data["service"] = self.get_service_info()
-        data["process"] = self.get_process_info()
-        data["system"] = self.get_system_info()
-        data.update(**kwargs)
-        return data
+    def _build_metadata(self):
+        return {
+            "service": self.get_service_info(),
+            "process": self.get_process_info(),
+            "system": self.get_system_info(),
+        }
 
     def _build_msg_for_logging(
         self, event_type, date=None, context=None, custom=None, stack=None, handled=True, **kwargs
@@ -471,10 +342,6 @@ class Client(object):
         else:
             context["custom"] = custom
 
-        # Run the data through processors
-        for processor in self.processors:
-            event_data = processor(self, event_data)
-
         # Make sure all data is coerced
         event_data = transform(event_data)
         if "exception" in event_data:
@@ -484,9 +351,15 @@ class Client(object):
 
         transaction = get_transaction()
         if transaction:
-            event_data["transaction"] = {"id": transaction.id}
+            if self.config.enable_distributed_tracing:
+                if transaction.trace_parent:
+                    event_data["trace_id"] = transaction.trace_parent.trace_id
+                event_data["parent_id"] = transaction.id
+                event_data["transaction_id"] = transaction.id
+            else:
+                event_data["transaction"] = {"id": transaction.id}
 
-        return self._build_msg({"errors": [event_data]})
+        return event_data
 
     def _filter_exception_type(self, data):
         exception = data.get("exception")
@@ -508,28 +381,6 @@ class Client(object):
                 self.logger.info("Ignored %s exception due to exception type filter", exc_name)
                 return True
         return False
-
-    def _get_log_message(self, data):
-        # decode message so we can show the actual event
-        try:
-            data = self.decode(data)
-        except Exception:
-            message = "<failed decoding data>"
-        else:
-            message = data.pop("message", "<no message value>")
-        return message
-
-    def _get_transport(self, parsed_url):
-        if hasattr(self._transport_class, "sync_transport") and is_master_process():
-            # when in the master process, always use SYNC mode. This avoids
-            # the danger of being forked into an inconsistent threading state
-            self.logger.info("Sending message synchronously while in master " "process. PID: %s", os.getpid())
-            return self._transport_class.sync_transport(parsed_url)
-        if parsed_url not in self._transports:
-            self._transports[parsed_url] = self._transport_class(
-                parsed_url, verify_server_cert=self.config.verify_server_cert
-            )
-        return self._transports[parsed_url]
 
     def _get_stack_info_for_trace(
         self,