+++ tests/handlers/logbook/logbook_tests.py
@@ -2,6 +2,7 @@ import logbook
 import pytest
 from logbook import LogRecord
 
+from elasticapm.conf.constants import ERROR
 from elasticapm.handlers.logbook import LogbookHandler
 
 
@@ -21,7 +22,7 @@ def test_logbook_logger_error_level(logbook_logger, logbook_handler):
         logbook_logger.error("This is a test error")
 
     assert len(logbook_handler.client.events) == 1
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert event["log"]["logger_name"] == __name__
     assert event["log"]["level"] == "error"
     assert event["log"]["message"] == "This is a test error"
@@ -35,7 +36,7 @@ def test_logger_warning_level(logbook_logger, logbook_handler):
     with logbook_handler.applicationbound():
         logbook_logger.warning("This is a test warning")
     assert len(logbook_handler.client.events) == 1
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert event["log"]["logger_name"] == __name__
     assert event["log"]["level"] == "warning"
     assert event["log"]["message"] == "This is a test warning"
@@ -51,7 +52,7 @@ def test_logger_without_stacktrace_config(logbook_logger, logbook_handler):
     with logbook_handler.applicationbound():
         logbook_logger.warning("This is a test warning")
 
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert "stacktrace" not in event["log"]
 
 
@@ -61,7 +62,7 @@ def test_logger_without_stacktrace_stack_false(logbook_logger, logbook_handler):
     with logbook_handler.applicationbound():
         logbook_logger.warning("This is a test warning", stack=False)
 
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert "stacktrace" not in event["log"]
 
 
@@ -69,7 +70,7 @@ def test_logger_with_extra(logbook_logger, logbook_handler):
     with logbook_handler.applicationbound():
         logbook_logger.info("This is a test info with a url", extra=dict(url="http://example.com"))
     assert len(logbook_handler.client.events) == 1
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert event["context"]["custom"]["url"] == "http://example.com"
     assert "stacktrace" in event["log"]
     assert "exception" not in event
@@ -85,7 +86,7 @@ def test_logger_with_exc_info(logbook_logger, logbook_handler):
             logbook_logger.info("This is a test info with an exception", exc_info=True)
 
     assert len(logbook_handler.client.events) == 1
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
 
     assert event["log"]["message"] == "This is a test info with an exception"
     assert "stacktrace" in event["log"]
@@ -101,7 +102,7 @@ def test_logger_param_message(logbook_logger, logbook_handler):
     with logbook_handler.applicationbound():
         logbook_logger.info("This is a test of %s", "args")
     assert len(logbook_handler.client.events) == 1
-    event = logbook_handler.client.events.pop(0)["errors"][0]
+    event = logbook_handler.client.events[ERROR][0]
     assert event["log"]["message"] == "This is a test of args"
     assert "stacktrace" in event["log"]
     assert "exception" not in event