+++ tests/transports/test_urllib3.py
@@ -19,26 +19,26 @@ responses = Responses("urllib3")
 
 def test_send(waiting_httpserver):
     waiting_httpserver.serve_content(code=202, content="", headers={"Location": "http://example.com/foo"})
-    transport = Transport(urlparse.urlparse(waiting_httpserver.url))
-    url = transport.send(compat.b("x"), {})
+    transport = Transport(waiting_httpserver.url)
+    url = transport.send(compat.b("x"))
     assert url == "http://example.com/foo"
 
 
 @responses.activate
 def test_timeout():
-    transport = Transport(urlparse.urlparse("http://localhost"))
+    transport = Transport("http://localhost", timeout=5)
     responses.add("POST", "/", status=202, body=MaxRetryError(None, None, reason=TimeoutError()))
     with pytest.raises(TransportException) as exc_info:
-        transport.send("x", {}, timeout=5)
+        transport.send("x")
     assert "timeout" in str(exc_info.value)
 
 
 def test_http_error(waiting_httpserver):
     waiting_httpserver.serve_content(code=418, content="I'm a teapot")
-    transport = Transport(urlparse.urlparse(waiting_httpserver.url))
+    transport = Transport(waiting_httpserver.url)
 
     with pytest.raises(TransportException) as exc_info:
-        transport.send("x", {})
+        transport.send("x")
     for val in (418, "I'm a teapot"):
         assert str(val) in str(exc_info.value)
 
@@ -46,28 +46,28 @@ def test_http_error(waiting_httpserver):
 @responses.activate
 def test_generic_error():
     url, status, message, body = ("http://localhost:9999", 418, "I'm a teapot", "Nothing")
-    transport = Transport(urlparse.urlparse(url))
+    transport = Transport(url)
     responses.add("POST", "/", status=status, body=Exception("Oopsie"))
     with pytest.raises(TransportException) as exc_info:
-        transport.send("x", {})
+        transport.send("x")
     assert "Oopsie" in str(exc_info.value)
 
 
 def test_http_proxy_environment_variable():
     with mock.patch.dict("os.environ", {"HTTP_PROXY": "http://example.com"}):
-        transport = Transport(urlparse.urlparse("http://localhost:9999"))
+        transport = Transport("http://localhost:9999")
         assert isinstance(transport.http, urllib3.ProxyManager)
 
 
 def test_https_proxy_environment_variable():
     with mock.patch.dict("os.environ", {"HTTPS_PROXY": "https://example.com"}):
-        transport = Transport(urlparse.urlparse("http://localhost:9999"))
+        transport = Transport("http://localhost:9999")
         assert isinstance(transport.http, urllib3.poolmanager.ProxyManager)
 
 
 def test_https_proxy_environment_variable_is_preferred():
     with mock.patch.dict("os.environ", {"HTTPS_PROXY": "https://example.com", "HTTP_PROXY": "http://example.com"}):
-        transport = Transport(urlparse.urlparse("http://localhost:9999"))
+        transport = Transport("http://localhost:9999")
         assert isinstance(transport.http, urllib3.poolmanager.ProxyManager)
         assert transport.http.proxy.scheme == "https"
 
@@ -79,11 +79,11 @@ def test_header_encodings():
     encoded bytestring, and explodes.
     """
     headers = {compat.text_type("X"): compat.text_type("V")}
-    transport = Transport(urlparse.urlparse("http://localhost:9999"))
+    transport = Transport("http://localhost:9999", headers=headers)
 
     with mock.patch("elasticapm.transport.http.urllib3.PoolManager.urlopen") as mock_urlopen:
         mock_urlopen.return_value = mock.Mock(status=202)
-        transport.send("", headers)
+        transport.send("")
     _, args, kwargs = mock_urlopen.mock_calls[0]
     if compat.PY2:
         assert isinstance(args[1], compat.binary_type)
@@ -94,15 +94,15 @@ def test_header_encodings():
 
 def test_ssl_verify_fails(waiting_httpsserver):
     waiting_httpsserver.serve_content(code=202, content="", headers={"Location": "http://example.com/foo"})
-    transport = Transport(compat.urlparse.urlparse(waiting_httpsserver.url))
+    transport = Transport(waiting_httpsserver.url)
     with pytest.raises(TransportException) as exc_info:
-        url = transport.send(compat.b("x"), {})
+        url = transport.send(compat.b("x"))
     assert "CERTIFICATE_VERIFY_FAILED" in str(exc_info)
 
 
 @pytest.mark.filterwarnings("ignore:Unverified HTTPS")
 def test_ssl_verify_disable(waiting_httpsserver):
     waiting_httpsserver.serve_content(code=202, content="", headers={"Location": "https://example.com/foo"})
-    transport = Transport(compat.urlparse.urlparse(waiting_httpsserver.url), verify_server_cert=False)
-    url = transport.send(compat.b("x"), {})
+    transport = Transport(waiting_httpsserver.url, verify_server_cert=False)
+    url = transport.send(compat.b("x"))
     assert url == "https://example.com/foo"